+++ src/main/java/br/com/fiap/fintech/repository/BankAccountRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BankAccountRepository extends JpaRepository<BankAccount, Integer> {
-    List<BankAccount> findByUserId(int userId);
+public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
+    List<BankAccount> findByUserId(Long userId);
 }
