+++ src/main/java/br/com/fiap/fintech/repository/BankRepository.java
@@ -3,6 +3,6 @@ package br.com.fiap.fintech.repository;
 import br.com.fiap.fintech.model.Bank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BankRepository extends JpaRepository<Bank, Integer> {
+public interface BankRepository extends JpaRepository<Bank, Long> {
 
 }
