+++ src/main/java/br/com/fiap/fintech/repository/BranchRepository.java
@@ -3,6 +3,6 @@ package br.com.fiap.fintech.repository;
 import br.com.fiap.fintech.model.Branch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BranchRepository extends JpaRepository<Branch, Integer> {
+public interface BranchRepository extends JpaRepository<Branch, Long> {
 
 }
