+++ src/main/java/br/com/fiap/fintech/repository/TransactionCategoryRepository.java
@@ -3,6 +3,6 @@ package br.com.fiap.fintech.repository;
 import br.com.fiap.fintech.model.TransactionCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionCategoryRepository extends JpaRepository<TransactionCategory, Integer> {
+public interface TransactionCategoryRepository extends JpaRepository<TransactionCategory, Long> {
 
 }
