+++ src/main/java/br/com/fiap/fintech/repository/TransactionTypeRepository.java
@@ -3,6 +3,6 @@ package br.com.fiap.fintech.repository;
 import br.com.fiap.fintech.model.TransactionType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionTypeRepository extends JpaRepository<TransactionType, Integer> {
+public interface TransactionTypeRepository extends JpaRepository<TransactionType, Long> {
 
 }
